feat: Fixed import of `useStore` in multiple files and added `createStoreProvider` function to impact-react
