+++ docs/.vitepress/theme/Playground.vue
@@ -23,7 +23,7 @@ async function loadSandpack(iframe: HTMLIFrameElement, code: string) {
         dependencies: {
           react: "18.2.0",
           "react-dom": "18.2.0",
-          "impact-react": "1.1.0",
+          "impact-react": "1.2.0",
         },
       }),
     },
@@ -71,7 +71,7 @@ root.render(
   // Optional options
   const options: ClientOptions = {
     showOpenInCodeSandbox: false,
-    bundlerURL: "https://sandpack-bundler.codesandbox.io"
+    //bundlerURL: "https://sandpack-bundler.codesandbox.io",
   };
 
   // Properly load and mount the bundler
