+++ docs/advanced/lists.md
@@ -2,9 +2,9 @@
 codeCaption: Working with lists
 code: |
   import { memo } from 'react'
-  import { store, signal, cleanup } from 'impact-react'
+  import { useStore, signal, cleanup, createStoreProvider } from 'impact-react'
 
-  const useStore = store(({ initialItems }) => {
+  function ListStore({ initialItems }) {
     // We create a signal for managing what to display in the list. It
     // only contains ids to items
     const listById = signal(initialItems.map((item) => item.id))
@@ -31,12 +31,15 @@ code: |
         return items[id].value
       }
     }
-  })
+  }
+
+  const useListStore = () => useStore(ListStore)
+  const ListStoreProvider = createStoreProvider(ListStore)
 
   // You memoize so that when the list changes the component does
   // not need to reconcile
   const Item = memo(function Item({ id }) {
-    const { getItemById } = useStore()
+    const { getItemById } = useListStore()
     // You consume the specific item directly from the context
     const item = getItemById(id)
 
@@ -48,16 +51,16 @@ code: |
   })
 
   function List() {
-  const { listById } = useStore()
+    const { listById } = useListStore()
 
-  return <ul>{listById.map((id) => <Item key={id} id={id} />)}</ul>
+    return <ul>{listById.map((id) => <Item key={id} id={id} />)}</ul>
   }
 
   export default function App() {
     return (
-      <useStore.Provider initialItems={[{ id: '123', title: "woop" }]}>
+      <ListStoreProvider initialItems={[{ id: '123', title: "woop" }]}>
         <List />
-      </useStore.Provider>
+      </ListStoreProvider>
     )
   }
 ---
@@ -67,9 +70,9 @@ code: |
 Creating observable lists in **Impact** is straight forward. It is just a signal with an array.
 
 ```ts
-import { store, signal } from 'impact-react'
+import { signal } from 'impact-react'
 
-const useStore = store(() => {
+function ListStore() {
   const list = signal([])
 
   return {
@@ -77,15 +80,15 @@ const useStore = store(() => {
       return list.value
     }
   }
-})
+}
 ```
 
 Since signal values are considered immutable, like in React, you update that list by:
 
 ```ts
-import { store, signal } from 'impact-react'
+import { signal } from 'impact-react'
 
-const useStore = store(() => {
+function ListStore() {
   const list = signal([])
 
   return {
@@ -96,17 +99,17 @@ const useStore = store(() => {
       list.value = [item, ...list.value]
     }
   }
-})
+}
 ```
 
 But sometimes lists needs to update complex objects within that list. In this scenario it can be a good idea to use something like [immer](https://immerjs.github.io/immer/):
 
 
 ```ts
-import { store, signal } from 'impact-react'
+import { signal } from 'impact-react'
 import { produce } from 'immer'
 
-const useStore = store(() => {
+function ListStore() {
   const list = signal([])
 
   return {
@@ -124,7 +127,7 @@ const useStore = store(() => {
       })
     }
   }
-})
+}
 ```
 
 Most lists can be managed this way, but you might have much bigger lists. Lists with sorting, filters and other states. You want to ensure maximum performance so that the component managing the list does not reconcile when items in the list updates and you want full control of what items are shown in the list at any moment.
