+++ docs/derived.md
@@ -7,9 +7,9 @@ outline: deep
 Derived reactive state. Will observe any signals or other derived in its callback. It lazily evaluates, which means when obseration triggers it only flags itself as dirty. The derived needs to be accessed to recaulcate its value.
 
 ```ts
-import { store, signal, derived } from 'impact-react'
+import { signal, derived } from 'impact-react'
 
-const useStore = store(() => {
+function MyStore() {
   const count = signal(0)
   const doubleCount = derived(() => count * 2)
 
@@ -21,5 +21,5 @@ const useStore = store(() => {
       return doubleCount.value
     }
   }
-})
+}
 ```
