+++ docs/effect.md
@@ -7,9 +7,9 @@ outline: deep
 Reactive effect. Will oberve and signals or derived in its callback. It runs immediately and will run again whenever obervation triggers. If an effect both observes and sets the same signal, the observation is ignored.
 
 ```ts
-import { store, effect, signal } from 'impact-react'
+import { effect, signal } from 'impact-react'
 
-const useStore = store(() => {
+function MyStore() {
   const count = signal(0)
 
   effect(() => {
@@ -21,6 +21,5 @@ const useStore = store(() => {
       return count.value
     }
   }
-})
-
+}
 ```
