+++ docs/learn/derived.md
@@ -1,9 +1,9 @@
 ---
 codeCaption: Cached derived signals
 code: |
-  import { store, signal, derived } from 'impact-react'
+  import { useStore, signal, derived } from 'impact-react'
 
-  const useStore = store(() => {
+  function CounterStore() {
     const count = signal(0)
     const enabled = signal(false)
     const multipliedCount = derived(() =>
@@ -28,10 +28,12 @@ code: |
         return multipliedCount.value
       }
     }
-  })
+  }
+
+  const useCounterStore = () => useStore(CounterStore)
 
   function Counter() {
-    const { count, increase } = useStore()
+    const { count, increase } = useCounterStore()
 
     return (
       <button onClick={increase}>
@@ -41,7 +43,7 @@ code: |
   }
 
   function Enabler() {
-    const { enabled, toggleEnabled } = useStore()
+    const { enabled, toggleEnabled } = useCounterStore()
 
     return (
       <button onClick={toggleEnabled}>
@@ -51,7 +53,7 @@ code: |
   }
 
   function Multiplier() {
-    const { multipliedCount } = useStore()
+    const { multipliedCount } = useCounterStore()
 
     return <h3>Multiplied: {multipliedCount}</h3>
   }
@@ -69,9 +71,9 @@ code: |
 
 # Derived
 
-Derived signals will calculate a value based on other signals and cache it. The benefit `derived` has over `useMemo` is that they do not immediately recalculate when a dependent signal changes, but rather flag itself as dirty. Only when the value is accessed it will recompute the value.
+Derived signals will calculate a value based on other signals and cache it. The benefit `derived` has over `useMemo` is that it does not immediately recalculate when a dependent signal changes, but rather flag itself as dirty. Only when the value is accessed it will recompute the value.
 
-Derived is consumed just like a plain signal, using the `.value` property, but you can not assign a value to a derived.
+Derived is consumed just like a plain signal, using the `.value` property. You can not assign a value to a derived.
 
 <ClientOnly>
   <Playground />
