+++ docs/learn/index.md
@@ -15,7 +15,6 @@ code: |
       </button>
     )
   }
-
 ---
 
 # State management
