+++ docs/learn/scoping.md
@@ -1,9 +1,9 @@
 ---
 codeCaption: Scoping stores
 code: |
-  import { store, signal, cleanup } from 'impact-react'
+  import { useStore, signal, createStoreProvider, cleanup } from 'impact-react'
 
-  const useStore = store(({ initialCount }) => {
+  function CounterStore({ initialCount }) {
     const count = signal(initialCount)
     const interval = setInterval(() => {
       count.value++
@@ -16,37 +16,34 @@ code: |
         return count.value
       }
     }
-  })
+  }
+
+  const CounterStoreProvider = createStoreProvider(CounterStore)
 
   function Counter() {
-    const { count } = useStore()
+    const { count } = useStore(CounterStore)
 
-    return (
-      <h2>
-        Count {count}
-      </h2>
-    )
+    return <h2>Count {count}</h2>
   }
 
   export default function App() {
     return (
-      <useStore.Provider initialCount={10}>
+      <CounterStoreProvider initialCount={10}>
         <Counter />
-      </useStore.Provider>
+      </CounterStoreProvider>
     )
   }
 ---
 
 # Scoping
 
-By default the stores are global, but you can scope them to specific component trees by using their provider.
+By default the stores are global, but you can scope them to specific component trees by using a provider.
 
-The hook used with the store has a `.Provider` property on it. Under the hood the store is exposed through a React context. The store is instantiated when the provider mounts and any `cleanup` is called when the provider unmounts.
+You create a provider for a store using the `createStoreProvider` function. The store is instantiated when the provider mounts and any `cleanup` is called when the provider unmounts.
 
 Scoping stores allows you to instantiate state management related to specific pages, features or even for each item in a list.
 
 Additionally the store can now receive props from the provider to initialise itself. This is especially useful to take advantage of modern React data fetching patterns, as you will see later.
-
 
 <ClientOnly>
  <Playground />
