+++ docs/signal.md
@@ -7,9 +7,9 @@ outline: deep
 Reactive state. The value is considered immutable and needs to be replaced. If replaced with the same value, the signal will not trigger.
 
 ```ts
-import { store, signal } from 'impact-react'
+import { signal } from 'impact-react'
 
-const useStore = store(() => {
+function MyStore() {
   const count = signal(0)
 
   return {
@@ -17,5 +17,5 @@ const useStore = store(() => {
       return count.value
     }
   }
-})
+}
 ```
