+++ docs/store.md
@@ -4,59 +4,34 @@ outline: deep
 
 # Store
 
-## store (declarative)
+## useStore
 
-A simple declarative store.
-
-```ts
-import { store } from 'impact-react'
-
-const useStore = store({
-  count: 0,
-  increase() {
-    this.count++
-  },
-  get doubleCount() {
-    return this.count * 2
-  }
-})
-```
-
-## store (composable)
-
-A composable store allowing for composition with other stores, a private scope, effects and cleanup.
+The hook the allows you to consume your store in a component or from other stores.
 
 ```ts
-import { store, signal, derived, effect } from 'impact-react'
+import { useStore, signal } from 'impact-react'
 
-const useStore = store(() => {
+function MyStore() {
   const count = signal()
-  const doubleCount = derived(() => count.value * 2)
-
-  effect(() => console.log(count.value))
 
   return {
     get count() {
       return count.value
-    },
-    get doubleCount() {
-      return doubleCount.value
-    },
-    increase() {
-      count.value++
     }
   }
-})
+}
+
+export const useMyStore = () => useStore(MyStore)
 ```
 
-## store.Provider
+## createStoreProvider
 
 A store can be scoped to a component tree. This allows for passing props to the store.
 
 ```tsx
-import { store, signal } from 'impact-react'
+import { store, signal, createStoreProvider } from 'impact-react'
 
-const useStore = store(({ initialCount }) => {
+function MyStore({ initialCount }) {
   const count = signal(initialCount)
 
   return {
@@ -67,10 +42,12 @@ const useStore = store(({ initialCount }) => {
       count.value++
     }
   }
-})
+}
+
+const MyStoreProvider = createStoreProvider(MyStore)
 
 function Counter() {
-  const { count, increase } = useStore()
+  const { count, increase } = useStore(MyStore)
 
   return (
     <button onClick={increase}>
@@ -81,35 +58,18 @@ function Counter() {
 
 function App() {
   return (
-    <useStore.Provider initialCount={10}>
+    <MyStoreProvider initialCount={10}>
       <Counter />
-    </useStore.Provider>
+    </MyStoreProvider>
   )
 }
 ```
 
-## store.provide
-
-Provide the store using a higher order component.
+Alternatively use the higher order component:
 
 ```tsx
-import { store, signal } from 'impact-react'
-
-const useStore = store(({ initialCount }) => {
-  const count = signal(initialCount)
-
-  return {
-    get count() {
-      return count.value
-    },
-    increase() {
-      count.value++
-    }
-  }
-})
-
-const Counter = useStore.provide(function Counter() {
-  const { count, increase } = useStore()
+const App = MyStoreProvider.provide(function App() {
+  const { count, increase } = useStore(MyStore)
 
   return (
     <button onClick={increase}>
@@ -117,8 +77,5 @@ const Counter = useStore.provide(function Counter() {
     </button>
   )
 })
-
-function App() {
-  return <Counter initialCount={10} />
-}
 ```
+
