+++ impact-react/package.json
@@ -1,6 +1,6 @@
 {
   "name": "impact-react",
-  "version": "1.1.1",
+  "version": "1.2.0",
   "description": "Reactive state management for React",
   "author": "Christian Alfoni <christianalfoni@gmail.com>",
   "license": "MIT",
